backend/requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-community==0.0.38
chromadb==0.4.22
sentence-transformers==2.3.1
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
